2.18 afternoon

+++ 05-dbpms-admin/src/main/java/com/cy/pj/sys/service/serviceImpl/SysRoleServiceImpl.java
@@ -1,15 +1,17 @@
 package com.cy.pj.sys.service.serviceImpl;
 
+import com.cy.pj.common.exception.ServiceException;
 import com.cy.pj.common.pojo.CheckBox;
-import com.cy.pj.sys.dao.SysMenuDao;
 import com.cy.pj.sys.dao.SysRoleDao;
 import com.cy.pj.sys.dao.SysRoleMenuDao;
 import com.cy.pj.sys.pojo.SysRole;
 import com.cy.pj.sys.service.SysRoleService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
 public class SysRoleServiceImpl implements SysRoleService {
     @Autowired
     private SysRoleDao sysRoleDao;
@@ -37,9 +39,10 @@ public class SysRoleServiceImpl implements SysRoleService {
 
     @Override
     public int updateRole(SysRole entity) {
-        //保存角色自身信息
+        //更新角色自身信息
         int rows = sysRoleDao.updateRole(entity);
-        //保存角色菜单关系数据
+        if (rows == 0) throw new ServiceException("记录可能已经不存在了");
+        //更新角色菜单关系数据(one2many数据更新一般是先删除原有关系，再添加新的关系)
         sysRoleMenuDao.deleteByRoleId(entity.getId());
         sysRoleMenuDao.insertRoleMenus(entity.getId(), entity.getMenuIds());
         return rows;
@@ -47,6 +50,6 @@ public class SysRoleServiceImpl implements SysRoleService {
 
     @Override
     public List<CheckBox> findCheckRoles() {
-        return null;
+        return sysRoleDao.selectCheckRoles();
     }
 }
